+++ CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 [Unreleased]
 ------------
+### Added
+- New box in the side bar with a table of all the languages a user is 
+learning. It displays the Level, Total XP and XP to Next Level for each 
+language.
+- Options to enable/disable this feature, and change the way the list is 
+sorted, either by Language, XP or, XP to Next Level. You can also choose 
+the reverse of any of these.
+
 ### Changed
 - Changed LTR/RTL detection to search all elements in question area for dir
 attribute. This makes the detection more flexible to different question